refactor: isolate card elements into their own component
